use-cases/time-series-forecasting/walmart-sales/ML/requirements.txt
neptune-client
matplotlib
seaborn
numpy
pandas
prophet
scikit-learn
xgboost
